Can send UDP packets
The system can send UDP packets, unfortunately they do not seem to be
picked up by the rest of the system (such as netcat), but are picked up
by WireShark. I do not yet know why.

There are a few problems that need resolving in the design, certain
layers should be made to be agnostic to IPv4/IPv6 and what Link layer is
being used, this will need redesigning later.